+++ PluginSdk/BasicRecon/SamplingMaskCreator.cpp
@@ -57,7 +57,7 @@ bool Yap::SamplingMaskCreator::Input(const wchar_t * port, IData * data)
 		double sample_percent = GetProperty<double>(L"SamplePercent");
 		double radius = GetProperty<double>(L"Radius");
 
-		auto mask = GenerateRandomMask(width, height, pow, sample_percent, radius);
+		auto mask = GenerateRandomMask(width, height, (float)pow, (float)sample_percent, (float)radius);
 		float * mask_buffer = nullptr;
 		try
 		{
